@@ -2,13 +2,14 @@
 #
 # SPDX-License-Identifier: MIT
 
+"""Main file for the package."""
+
 import importlib.metadata
 import json
-import os
 import textwrap
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Union
+from typing import Any
 
 import packaging.version
 from docutils import nodes
@@ -20,18 +21,26 @@ INDENT = " " * 4
 
 @dataclass
 class Config:
-    """
-    Holds the configuration data for the extension
-    """
+    """Holds the configuration data for the extension."""
 
     datatables_version: str
     datatables_class: str
-    datatables_options: Union[dict, str]
+    datatables_options: dict | str
     datatables_js: str
     datatables_css: str
 
 
 def get_config(app: Sphinx) -> Config:
+    """
+    Return the configurable options.
+
+    Args:
+        app (Sphinx): Sphinx app
+
+    Returns:
+        Config: Dataclass for the config options
+
+    """
     return Config(
         datatables_version=app.config.datatables_version,
         datatables_class=app.config.datatables_class,
@@ -43,33 +52,35 @@ def get_config(app: Sphinx) -> Config:
 
 def add_datatables_scripts(
     app: Sphinx,
-    pagename: str,
-    templatename: str,
-    context: dict,
-    doctree: nodes.document,
+    _pagename: str,
+    _templatename: str,
+    _context: dict,
+    _doctree: nodes.document,
 ) -> None:
-    """
-    Add the scripts to enable Datatables
-    """
-
+    """Add the scripts to enable Datatables."""
     config = get_config(app)
 
     # Set up jQuery first, to verify it is available and gracefully output an error
     try:
         app.setup_extension("sphinxcontrib.jquery")
     except ExtensionError:
-        raise ExtensionError(
-            "sphinxcontrib.jquery is required for sphinx-datatables to work. "
-            "Please add it to your extensions in conf.py."
+        msg = (
+            "sphinxcontrib.jquery is required for sphinx-datatables to work.",
+            "Please add it to your extensions in conf.py.",
         )
+        raise ExtensionError(
+            msg,
+        ) from None
 
     datetables_version_str = config.datatables_version
     if packaging.version.parse(datetables_version_str) < packaging.version.parse(
-        "2.0.0"
+        "2.0.0",
     ):
         datatables_js = f"https://cdn.datatables.net/{datetables_version_str}/js/jquery.dataTables.min.js"
         datatables_css = f"https://cdn.datatables.net/{datetables_version_str}/css/jquery.dataTables.min.css"
-    else:  # this is for DataTables 2.0.0 and above, only the minified version is available (jQuery is not included)
+    else:
+        # for DataTables 2.0.0 and above, only the minified version is available
+        # and jQuery is not included
         datatables_js = f"https://cdn.datatables.net/v/dt/dt-{datetables_version_str}/datatables.min.js"
         datatables_css = f"https://cdn.datatables.net/v/dt/dt-{datetables_version_str}/datatables.min.css"
 
@@ -84,11 +95,12 @@ def add_datatables_scripts(
     app.add_js_file("activate_datatables.js")
 
 
-def datatables_options_to_js(options: Union[dict, str], indent: str) -> str:
+def datatables_options_to_js(options: dict | str, indent: str) -> str:
     """
-    Convert a Python nested dictionary to a valid JS dictionary object as a string
-    or the string itself if it's not a dict, but indented.
-    Appends a comma at the end if not already present.
+    Convert a Python nested dictionary to a valid JS dictionary object as a str.
+
+    If it's a string already, dedent and return. It will also append a comma at
+    the end if not already present.
     """
     if isinstance(options, dict):
         obj = json.dumps(options, indent=INDENT)
@@ -103,46 +115,52 @@ def datatables_options_to_js(options: Union[dict, str], indent: str) -> str:
 
 def create_datatables_js(
     datatables_class: str,
-    datatables_options: Union[dict, str],
+    datatables_options: dict | str,
     datatables_version: str,
 ) -> str:
-    """
-    Create the JS file to activate datatables
-    """
-    custom_file = str(
-        Path(__file__).parent.joinpath("activate_datatables.js").absolute()
-    )
-    with open(custom_file + ".in", "r") as template:
+    """Create the JS file to activate datatables."""
+    custom_file = Path(__file__).parent.joinpath("activate_datatables.js.in").absolute()
+    with custom_file.open() as template:
         contents = template.read()
         contents = contents.replace(r"${datatables_class}", datatables_class)
         datatables_options = datatables_options_to_js(datatables_options, INDENT * 2)
         datatables_options = datatables_options.replace(
-            r"${datatables_version}", datatables_version
+            r"${datatables_version}",
+            datatables_version,
         )
-        contents = contents.replace(r"${datatables_options}", datatables_options)
-    return contents
+        return contents.replace(r"${datatables_options}", datatables_options)
 
 
-def finish(app: Sphinx, exception) -> None:
+def finish(app: Sphinx, _exception: Exception | None) -> None:
+    """
+    Save the assets to the static directory.
+
+    This function is called as the build finishes.
+
+    Args:
+        app (Sphinx): Sphinx app
+        _exception (Exception | None): Any exceptions from the build
+
+    """
     config = get_config(app)
     datatables_config_contents = create_datatables_js(
         config.datatables_class,
         config.datatables_options,
         config.datatables_version,
     )
-    asset_file = os.path.join(app.builder.outdir, "_static/activate_datatables.js")
-    with open(asset_file, "w+") as f:
+    asset_file = Path(app.builder.outdir) / "_static/activate_datatables.js"
+    with asset_file.open("w+") as f:
         f.write(datatables_config_contents)
 
 
 def setup(app: Sphinx) -> dict[str, Any]:
     """
-    Setup the extension
+    Set up the extension.
 
     Args:
         app (Sphinx): Sphinx app
-    """
 
+    """
     app.add_config_value("datatables_version", "2.3.5", "html", str)
     app.add_config_value("datatables_class", "sphinx-datatable", "html", str)
     app.add_config_value("datatables_options", {}, "html", [dict, str])
